requirements.txt
altair==5.5.0
altex==0.2.0
annotated-types==0.7.0
anyio==4.10.0
asn1crypto==1.5.1
attrs==25.3.0
Authlib==1.6.5
beautifulsoup4==4.14.2
blinker==1.9.0
boto3==1.40.60
botocore==1.40.60
CacheControl==0.14.3
cachetools==5.5.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
cloudpickle==3.1.1
contourpy==1.3.2
cryptography==46.0.1
cycler==0.12.1
entrypoints==0.4
exceptiongroup==1.3.0
Faker==37.12.0
favicon==0.7.0
filelock==3.20.0
firebase_admin==7.1.0
fonttools==4.60.1
gcloud==0.18.3
gitdb==4.0.12
GitPython==3.1.45
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.4.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.75.0
grpcio-status==1.75.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
htbuilder==0.9.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
Jinja2==3.1.6
jmespath==1.0.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jwcrypto==1.5.6
kiwisolver==1.4.9
lxml==6.0.2
Markdown==3.9
markdownlit==0.0.7
MarkupSafe==3.0.2
matplotlib==3.10.7
msgpack==1.1.1
narwhals==2.5.0
numpy==2.2.6
oauth2client==4.1.3
ollama==0.6.0
packaging==25.0
pandas==2.3.2
pillow==11.3.0
platformdirs==4.5.0
plotly==6.3.1
prometheus_client==0.23.1
proto-plus==1.26.1
protobuf==6.31.1
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.12.3
pydantic_core==2.41.4
pydeck==0.9.1
PyJWT==2.10.1
pymdown-extensions==10.16.1
pyOpenSSL==25.3.0
pyparsing==3.2.5
Pyrebase4==4.8.0
python-dateutil==2.9.0.post0
python-jwt==4.1.0
pytz==2025.2
PyYAML==6.0.3
referencing==0.36.2
requests==2.32.5
requests-toolbelt==0.10.1
rpds-py==0.27.1
rsa==4.9.1
s3transfer==0.14.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
snowflake-connector-python==4.0.0
snowflake-snowpark-python==1.42.0
sortedcontainers==2.4.0
soupsieve==2.8
st-annotated-text==4.0.2
st-theme==1.2.3
streamlit==1.49.1
streamlit-avatar==0.1.3
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.2
streamlit-embedcode==0.1.2
streamlit-extras==0.7.8
streamlit-float==0.3.5
streamlit-image-coordinates==0.4.0
streamlit-keyup==0.3.0
streamlit-notify==0.3.1
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
streamlit_faker==0.0.4
tenacity==9.1.2
toml==0.10.2
tomlkit==0.13.3
tornado==6.5.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==1.26.20
validators==0.35.0
setuptools